[lsp] use language specific proposal acceptor, added SerializableEObjectDescriptionProvider hook
