+++ org.eclipse.xtext.ide/xtend-gen/org/eclipse/xtext/ide/server/LanguageServerImpl.java
@@ -74,10 +74,8 @@ import io.typefox.lsapi.services.WindowService;
 import io.typefox.lsapi.services.WorkspaceService;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeSet;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import org.eclipse.emf.common.util.URI;
@@ -85,7 +83,7 @@ import org.eclipse.xtend.lib.annotations.Accessors;
 import org.eclipse.xtend.lib.annotations.FinalFieldsConstructor;
 import org.eclipse.xtext.diagnostics.Severity;
 import org.eclipse.xtext.ide.editor.contentassist.ContentAssistEntry;
-import org.eclipse.xtext.ide.editor.contentassist.IIdeContentProposalAcceptor;
+import org.eclipse.xtext.ide.editor.contentassist.IdeContentProposalAcceptor;
 import org.eclipse.xtext.ide.server.Document;
 import org.eclipse.xtext.ide.server.UriExtensions;
 import org.eclipse.xtext.ide.server.WorkspaceManager;
@@ -112,7 +110,6 @@ import org.eclipse.xtext.xbase.lib.Functions.Function2;
 import org.eclipse.xtext.xbase.lib.IterableExtensions;
 import org.eclipse.xtext.xbase.lib.ListExtensions;
 import org.eclipse.xtext.xbase.lib.ObjectExtensions;
-import org.eclipse.xtext.xbase.lib.Pair;
 import org.eclipse.xtext.xbase.lib.Procedures.Procedure1;
 import org.eclipse.xtext.xbase.lib.Procedures.Procedure2;
 import org.eclipse.xtext.xbase.lib.Pure;
@@ -458,63 +455,8 @@ public class LanguageServerImpl implements LanguageServer, WorkspaceService, Win
       }
       final Function2<Document, XtextResource, ArrayList<CompletionItemImpl>> _function_1 = (Document document, XtextResource resource) -> {
         try {
-          final Comparator<Pair<Integer, ContentAssistEntry>> _function_2 = (Pair<Integer, ContentAssistEntry> p1, Pair<Integer, ContentAssistEntry> p2) -> {
-            Integer _key = p2.getKey();
-            Integer _key_1 = p1.getKey();
-            final int prioResult = _key.compareTo(_key_1);
-            if ((prioResult != 0)) {
-              return prioResult;
-            }
-            String _elvis = null;
-            ContentAssistEntry _value = p1.getValue();
-            String _label = _value.getLabel();
-            if (_label != null) {
-              _elvis = _label;
-            } else {
-              ContentAssistEntry _value_1 = p1.getValue();
-              String _proposal = _value_1.getProposal();
-              _elvis = _proposal;
-            }
-            final String s1 = _elvis;
-            String _elvis_1 = null;
-            ContentAssistEntry _value_2 = p2.getValue();
-            String _label_1 = _value_2.getLabel();
-            if (_label_1 != null) {
-              _elvis_1 = _label_1;
-            } else {
-              ContentAssistEntry _value_3 = p2.getValue();
-              String _proposal_1 = _value_3.getProposal();
-              _elvis_1 = _proposal_1;
-            }
-            final String s2 = _elvis_1;
-            final int ignoreCase = s1.compareToIgnoreCase(s2);
-            if ((ignoreCase == 0)) {
-              return s1.compareTo(s2);
-            }
-            return ignoreCase;
-          };
-          final TreeSet<Pair<Integer, ContentAssistEntry>> entries = new TreeSet<Pair<Integer, ContentAssistEntry>>(_function_2);
-          final IIdeContentProposalAcceptor acceptor = new IIdeContentProposalAcceptor() {
-            @Override
-            public void accept(final ContentAssistEntry entry, final int priority) {
-              if ((entry != null)) {
-                String _proposal = entry.getProposal();
-                boolean _tripleEquals = (_proposal == null);
-                if (_tripleEquals) {
-                  throw new IllegalArgumentException("proposal must not be null.");
-                }
-                Pair<Integer, ContentAssistEntry> _mappedTo = Pair.<Integer, ContentAssistEntry>of(Integer.valueOf(priority), entry);
-                entries.add(_mappedTo);
-              }
-              LanguageServerImpl.this.operationCanceledManager.checkCanceled(cancelIndicator);
-            }
-            
-            @Override
-            public boolean canAcceptMoreProposals() {
-              int _size = entries.size();
-              return (_size < 100);
-            }
-          };
+          IResourceServiceProvider _resourceServiceProvider = resource.getResourceServiceProvider();
+          final IdeContentProposalAcceptor acceptor = _resourceServiceProvider.<IdeContentProposalAcceptor>get(IdeContentProposalAcceptor.class);
           Position _position = params.getPosition();
           final int caretOffset = document.getOffSet(_position);
           Position _position_1 = params.getPosition();
@@ -537,13 +479,13 @@ public class LanguageServerImpl implements LanguageServer, WorkspaceService, Win
             }
           }
           final ArrayList<CompletionItemImpl> completionItems = CollectionLiterals.<CompletionItemImpl>newArrayList();
-          final Procedure2<Pair<Integer, ContentAssistEntry>, Integer> _function_3 = (Pair<Integer, ContentAssistEntry> it, Integer idx) -> {
-            ContentAssistEntry _value = it.getValue();
-            final CompletionItemImpl item = this.toCompletionItem(_value, caretOffset, caretPosition, document);
+          Iterable<ContentAssistEntry> _entries = acceptor.getEntries();
+          final Procedure2<ContentAssistEntry, Integer> _function_2 = (ContentAssistEntry it, Integer idx) -> {
+            final CompletionItemImpl item = this.toCompletionItem(it, caretOffset, caretPosition, document);
             item.setSortText(("" + idx));
             completionItems.add(item);
           };
-          IterableExtensions.<Pair<Integer, ContentAssistEntry>>forEach(entries, _function_3);
+          IterableExtensions.<ContentAssistEntry>forEach(_entries, _function_2);
           return completionItems;
         } catch (Throwable _e) {
           throw Exceptions.sneakyThrow(_e);
