+++ org.eclipse.xtext.tests/xtend-gen/org/eclipse/xtext/resource/persistence/SerializableResourceDescriptionTest.java
@@ -75,12 +75,12 @@ public class SerializableResourceDescriptionTest {
           URI _appendFragment = uri.appendFragment("baz");
           it_1.setEObjectURI(_appendFragment);
           QualifiedName _create = QualifiedName.create("foo", "baz");
-          it_1.setQualifiedName(_create);
+          it_1.qualifiedName = _create;
           EClass _eAttribute = EcorePackage.eINSTANCE.getEAttribute();
           it_1.setEClass(_eAttribute);
           Pair<String, String> _mappedTo = Pair.<String, String>of("myKey", "myValue");
           HashMap<String, String> _newHashMap = CollectionLiterals.<String, String>newHashMap(_mappedTo);
-          it_1.setUserData(_newHashMap);
+          it_1.userData = _newHashMap;
         };
         SerializableEObjectDescription _doubleArrow_2 = ObjectExtensions.<SerializableEObjectDescription>operator_doubleArrow(_serializableEObjectDescription, _function_3);
         it.setDescriptions(Collections.<SerializableEObjectDescription>unmodifiableList(CollectionLiterals.<SerializableEObjectDescription>newArrayList(_doubleArrow_2)));
@@ -161,12 +161,8 @@ public class SerializableResourceDescriptionTest {
         QualifiedName _name = beforeDesc.getName();
         QualifiedName _name_1 = afterDesc.getName();
         Assert.assertEquals(_name, _name_1);
-        QualifiedName _qualifiedName = beforeDesc.getQualifiedName();
-        QualifiedName _qualifiedName_1 = afterDesc.getQualifiedName();
-        Assert.assertEquals(_qualifiedName, _qualifiedName_1);
-        HashMap<String, String> _userData = beforeDesc.getUserData();
-        HashMap<String, String> _userData_1 = afterDesc.getUserData();
-        Assert.assertEquals(_userData, _userData_1);
+        Assert.assertEquals(beforeDesc.qualifiedName, afterDesc.qualifiedName);
+        Assert.assertEquals(beforeDesc.userData, afterDesc.userData);
         URI _eObjectURI = beforeDesc.getEObjectURI();
         URI _eObjectURI_1 = afterDesc.getEObjectURI();
         Assert.assertEquals(_eObjectURI, _eObjectURI_1);
@@ -244,12 +240,12 @@ public class SerializableResourceDescriptionTest {
           URI _appendFragment = uri.appendFragment("baz");
           it_1.setEObjectURI(_appendFragment);
           QualifiedName _create = QualifiedName.create("foo", "baz");
-          it_1.setQualifiedName(_create);
+          it_1.qualifiedName = _create;
           EClass _eAttribute = EcorePackage.eINSTANCE.getEAttribute();
           it_1.setEClass(_eAttribute);
           Pair<String, String> _mappedTo = Pair.<String, String>of("myKey", "myValue");
           HashMap<String, String> _newHashMap = CollectionLiterals.<String, String>newHashMap(_mappedTo);
-          it_1.setUserData(_newHashMap);
+          it_1.userData = _newHashMap;
         };
         SerializableEObjectDescription _doubleArrow_5 = ObjectExtensions.<SerializableEObjectDescription>operator_doubleArrow(_serializableEObjectDescription, _function_6);
         it.setDescriptions(Collections.<SerializableEObjectDescription>unmodifiableList(CollectionLiterals.<SerializableEObjectDescription>newArrayList(_doubleArrow_5)));
